Conflicts:
	Duang/WebContent/page/home/left.html

+++ Duang/src/org/duang/action/sys/SysRoleAction.java
@@ -22,9 +22,12 @@ import org.duang.common.logger.LoggerUtils;
 import org.duang.entity.SysPower;
 import org.duang.entity.SysRole;
 import org.duang.entity.SysRolePower;
+import org.duang.entity.SysUser;
 import org.duang.service.SysPowerService;
 import org.duang.service.SysRoleService;
+import org.duang.service.SysUserService;
 import org.duang.util.DataUtils;
+import org.hibernate.criterion.Order;
 import org.springframework.context.annotation.Scope;
 import org.springframework.context.annotation.ScopedProxyMode;
 
@@ -56,6 +59,7 @@ public class SysRoleAction extends BaseAction<SysRole>{
 
 	private SysRoleService service;
 	private SysPowerService powerService;
+	private SysUserService sysUserService;
 	@Resource(name="sysroleserviceimpl")
 	public void setService(SysRoleService service) {
 		this.service = service;
@@ -64,7 +68,10 @@ public class SysRoleAction extends BaseAction<SysRole>{
 	public void setPowerService(SysPowerService powerService) {
 		this.powerService = powerService;
 	}
-
+	@Resource
+	public void setSysUserService(SysUserService sysUserService) {
+		this.sysUserService = sysUserService;
+	}
 	/**   
 	 * 跳转到角色管理界面
 	 * @Title: showRole   
@@ -78,6 +85,36 @@ public class SysRoleAction extends BaseAction<SysRole>{
 	public String showRole() {
 		return ResultPath.LIST;
 	}
+
+
+	/**   
+	 * 查询角色下拉列表
+	 * @Title: queryRoleList   
+	 * @Description: TODO(这里用一句话描述这个方法的作用)   
+	 * @param:   
+	 * @author 白攀    
+	 * @date 2016年7月28日 下午1:54:00
+	 * @return: void      
+	 * @throws   
+	 */  
+	public void queryRoleList() {
+		String json = "";
+		try {
+			List<SysRole> list = service.queryAllEntity(Order.desc("optionTime"));
+			for(SysRole role : list){
+				Map<String,Object> map = new HashMap<String,Object>();
+				map.put("roleId", role.getId());
+				map.put("roleName", role.getRoleName());
+				listMap.add(map);
+			}
+			json = JSONArray.fromObject(listMap).toString();
+		} catch (Exception e) {
+			e.printStackTrace();
+		} finally {
+			printJsonResult(json);
+		}
+	}
+
 
 	/**   
 	 * 页面跳转
@@ -292,10 +329,16 @@ public class SysRoleAction extends BaseAction<SysRole>{
 		try {
 			String sysRoleId = super.getRequest().getParameter("sysRoleId");
 			if (DataUtils.notEmpty(sysRoleId)) {
-				if (service.deleteEntity(sysRoleId)) {//删除中间表数据 并且 删除角色表数据
-					jsonObject.put("success", true);
-				}else {
+				List<SysUser> users = sysUserService.queryEntity("sysRole.id", sysRoleId, null, null);
+				if (users!=null && users.size()>0) {
 					jsonObject.put("success", false);
+					jsonObject.put("msg", "该角色拥有已关联用户，禁止删除");
+				}else {
+					if (service.deleteEntity(sysRoleId)) {//删除中间表数据 并且 删除角色表数据
+						jsonObject.put("success", true);
+					}else {
+						jsonObject.put("success", false);
+					}
 				}
 			}else{
 				jsonObject.put("success", false);
